# perf(expr): prebuild the AC for `to_char`